BOAC-3146, remove noteTemplates from mixin/UserMetadata; retire store/modules/note.ts
