+++ src/api/note-templates.ts
@@ -25,7 +25,7 @@ export function createNoteTemplate(
   const data = {title, subject, body, topics};
   _.each(attachments || [], (attachment, index) => data[`attachment[${index}]`] = attachment);
   return utils.postMultipartFormData('/api/note_template/create', data).then(template => {
-    store.dispatch('note/onCreateTemplate', template);
+    store.dispatch('noteEditSession/onCreateTemplate', template);
     store.dispatch('user/gaNoteTemplateEvent', {
       id: template.id,
       label: `Advisor ${store.getters['user/uid']} created a note template`,
@@ -38,14 +38,14 @@ export function createNoteTemplate(
 export function deleteNoteTemplate(templateId: number) {
   return axios
     .delete(`${utils.apiBaseUrl()}/api/note_template/delete/${templateId}`)
-    .then(() => store.dispatch('note/onDeleteTemplate', templateId));
+    .then(() => store.dispatch('noteEditSession/onDeleteTemplate', templateId));
 }
 
 export function renameNoteTemplate(noteTemplateId: number, title: string) {
   const data = {id: noteTemplateId, title: title};
   return axios.post(`${utils.apiBaseUrl()}/api/note_template/rename`, data).then(response => {
     const template = response.data;
-    store.dispatch('note/onUpdateTemplate', template);
+    store.dispatch('noteEditSession/onUpdateTemplate', template);
     store.dispatch('user/gaNoteTemplateEvent', {
       id: noteTemplateId,
       name: `Advisor ${store.getters['user/uid']} renamed a note template`,
@@ -72,7 +72,7 @@ export function updateNoteTemplate(
   };
   _.each(newAttachments || [], (attachment, index) => data[`attachment[${index}]`] = attachment);
   return utils.postMultipartFormData('/api/note_template/update', data).then(template => {
-    store.dispatch('note/onUpdateTemplate', template);
+    store.dispatch('noteEditSession/onUpdateTemplate', template);
     store.dispatch('user/gaNoteTemplateEvent', {
       id: noteTemplateId,
       name: `Advisor ${store.getters['user/uid']} updated a note template`,
